requirements.txt
PyQt6==6.7.0
qasync==0.27.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.0
